+++ jslint.mjs
@@ -2418,7 +2418,7 @@ function jslint_phase2_lex(state) {
 
 // PHASE 2. Lex <line_list> into <token_list>.
 
-    let {
+    const {
         artifact,
         directive_list,
         global_dict,
@@ -4175,16 +4175,20 @@ import moduleHttps from "https";
             break;
         case ")":
         case "]":
-            opener_popped = opener_stack.pop();
-            if (noop(
-                id === ")"
-                ? opener_popped.id !== "("
-                : opener_popped.id !== "["
-            )) {
+
+// PR-503 - Fix jslint crashing before warning about dangling ')' or ']'.
+
+            opener_popped = opener_stack.pop() || empty();
+            if (
+                (id === ")" && opener_popped.id !== "(")
+                || (id === "]" && opener_popped.id !== "[")
+            ) {
 
 // test_cause:
+// [")", "token_create", "unexpected_a", ")", 1]
 // [";(]", "token_create", "unexpected_a", "]", 3]
 // [";[)", "token_create", "unexpected_a", ")", 3]
+// ["]", "token_create", "unexpected_a", "]", 1]
 
                 return stop("unexpected_a", the_token);
             }
@@ -4266,8 +4270,7 @@ function jslint_phase3_parse(state) {
 
 // Specialized tokens may have additional properties.
 
-    let anon = "anonymous";     // The guessed name for anonymous functions.
-    let {
+    const {
         artifact,
         catch_list,
         catch_stack,
@@ -4288,6 +4291,7 @@ function jslint_phase3_parse(state) {
         warn,
         warn_at
     } = state;
+    let anon = "anonymous";     // The guessed name for anonymous functions.
     let catchage = catch_stack[0];      // The current catch-block.
     let functionage = token_global;     // The current function.
     let mode_var;               // "var" if using var; "let" if using let.
@@ -7818,7 +7822,7 @@ function jslint_phase4_walk(state) {
 //          recursive traversal. Each node may be processed on the way down
 //          (preaction) and on the way up (postaction).
 
-    let {
+    const {
         artifact,
         catch_stack,
         function_stack,
@@ -9152,7 +9156,7 @@ function jslint_phase5_whitage(state) {
 
 // PHASE 5. Check whitespace between tokens in <token_list>.
 
-    let {
+    const {
         artifact,
         catch_list,
         function_list,